split video source and partial wip on cameravideoSource
