+++ Runtime/Scripts/CameraVideoSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a687aca3b53dfe64fa4e6061a03d4f8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
