+++ Runtime/Scripts/RtcVideoSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 132c76bbe27f1e2419dcad5c7699c3e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
