+++ Runtime/Scripts/ScreenVideoSource.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b3995117cdee64fb3b81d5cf568e03c6
+guid: c41c73a604498c64da7a0016babb6090
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
