+++ Runtime/Scripts/TextureVideoSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a21dae64441674a478fe4ec708b9b059
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
